+++ src/main/java/webserver/http/HttpHeaders.java
@@ -1,4 +1,4 @@
-package webserver.http.request;
+package webserver.http;
 
 import utils.Parser;
